+++ jdk10/src/test/java/java10/optional/OptionalTest.java
@@ -1,0 +1,20 @@
+package java10.optional;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+@DisplayName("Optional")
+class OptionalTest {
+    @Test
+    void shouldThrowANoSuchElementException() {
+        // given
+        Optional<Object> optional = Optional.empty();
+        // expect
+        assertThrows(NoSuchElementException.class, optional::orElseThrow);
+    }
+}
