+++ jdk10/src/test/java/java10/var/VarTest.java
